Rename auth_bloc -> auth_navigation_bloc
