+++ lib/view/onboarding/auth_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:mobile/view/onboarding/auth_bloc.dart';
+import 'package:mobile/view/onboarding/auth_navigation_bloc.dart';
 import 'package:mobile/view/onboarding/login_screen.dart';
 import 'package:mobile/view/onboarding/signup_screen.dart';
